% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{loss_sum_of_squares}
\alias{loss_sum_of_squares}
\alias{loss_root_mean_square}
\title{Loss Functions}
\usage{
loss_sum_of_squares(observed, predicted)
}
\arguments{
\item{observed}{true observed lables}

\item{predicted}{scores predicted with a model}
}
\value{
numeric model fit assessment. The lower - the better is the model
}
\description{
Loss functions that can be used for model comparions ane feature importance estimation
}
