% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mi2.R
\name{theme_mi2}
\alias{theme_mi2}
\title{MI^2 Theme}
\usage{
theme_mi2()
}
\value{
theme object that can be added to ggplot2 plots
}
\description{
Themem oject. In order to avoid an additional dependency from ggplot2, we have here a raw vaerion of an theme object
}
