% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{yhat}
\alias{yhat}
\alias{yhat.lm}
\alias{yhat.randomForest}
\alias{yhat.default}
\title{Wrapper over the predict function}
\usage{
yhat(X.model, newdata, ...)

\method{yhat}{lm}(X.model, newdata, ...)

\method{yhat}{randomForest}(X.model, newdata, ...)

\method{yhat}{default}(X.model, newdata, ...)
}
\arguments{
\item{X.model}{object - a model to be explained}

\item{newdata}{data.frame or matrix - observations for prediction}

\item{...}{other parameters that will be passed to the `predict` function}
}
\value{
An numeric matrix of predictions. Can have more than one collumn.
}
\description{
This function is just a wrapper over the predict function.
It sets different default parameters for models from different classes.
For example: for classification random Forest is forces the output to be probabilities not classes itself.
}
\details{
We use the `yhat` name instead of `predict` since we need different defaults that the one set in the `predict` function.
}
